Finally got the ppms working
Looking forward to getting this working with `iter()` and `map()`, but I
was having trouble figuring out how to aggregate the strings with that
approach. As I get better with the basics of Rust I can fix some of this
up, but for now I want to get the ray tracer doing something interesting
enough that I'll stick with this project.